requirements.txt
streamlit
python-dotenv
langchain-core
langchain-mcp-adapters
langchain-google-genai
google-generativeai
